fixed 3 scripts to read from parameters.ini: backupDbs.sh, first_time_setup.sh, and updateFeedbackTables.sh

+++ misc_scripts/backupDbs.sh
@@ -1,14 +1,33 @@
 #!/bin/bash
-#TODO: Get db host, username, pass, schema from parameters.ini
 
-DB_HOST="localhost"
-PROD_DB_NAME="campsite"
-ACL_DB_NAME="campsite_acl"
-DB_USER="root"
-DB_PASS="sqladmin"
+SCRIPTPATH="$( cd "$(dirname "$0")" ; pwd -P )"
+if [ ${SCRIPTPATH##*/} != "misc_scripts" ]; then
+    echo "Fatal error: This script must be placed in the misc_scripts directory of campsite";
+    exit 1
+fi
 
-mysqldump -h$DB_HOST -u$DB_USER -p$DB_PASS $PROD_DB_NAME > prodDbBackup.sql
-mysqldump -h$DB_HOST -u$DB_USER -p$DB_PASS $ACL_DB_NAME > aclDbBackup.sql 
+INI_FILE_PATH="${SCRIPTPATH}/../app/config/parameters.ini";
+if [ ! -e $INI_FILE_PATH ]; then
+    echo "Fatal error: app/config/parameters.ini file not found."
+    exit 1
+fi
+
+# Get db username, pass, schema from parameters.ini
+DB_USER=$( sed -n 's/^ *database_user *= *\([^ ]*.*\)/\1/p' < ${INI_FILE_PATH} )
+DB_PASS=$( sed -n 's/^ *database_password *= *\([^ ]*.*\)/\1/p' < ${INI_FILE_PATH} )
+DB_HOST=$( sed -n 's/^ *database_host *= *\([^ ]*.*\)/\1/p' < ${INI_FILE_PATH} )
+PROD_DB_NAME=$( sed -n 's/^ *database_name *= *\([^ ]*.*\)/\1/p' < ${INI_FILE_PATH} )
+ACL_DB_NAME=$( sed -n 's/^ *acl_database_name *= *\([^ ]*.*\)/\1/p' < ${INI_FILE_PATH} )
+DB_PORT=$( sed -n 's/^ *database_port *= *\([^ ]*.*\)/\1/p' < ${INI_FILE_PATH} )
+
+if [ -z "$DB_USER" ] || [ -z "$DB_PASS" ] || [ -z "$DB_HOST" ] || [ -z "$DB_PORT" ] || [ -z "$PROD_DB_NAME" ] || [ -z "$ACL_DB_NAME" ]; then
+    echo "Fatal error: One of the following database properties in parameters.ini is empty:"
+    echo "   database_user, database_password, database_host, database_name, acl_database_name, database_port";
+    exit 1
+fi
+
+mysqldump -P$DB_PORT -h$DB_HOST -u$DB_USER -p$DB_PASS $PROD_DB_NAME > prodDbBackup.sql
+mysqldump -P$DB_PORT -h$DB_HOST -u$DB_USER -p$DB_PASS $ACL_DB_NAME > aclDbBackup.sql 
 
 tar -cvzf dbBackups.tar.gz prodDbBackup.sql aclDbBackup.sql
 rm prodDbBackup.sql aclDbBackup.sql
